requirements.txt
django==1.10
djangorestframework==3.4
markdown
django-oauth-toolkit